Add a caching wrapper around S3DB
Introduce a caching wrapper around the S3 DB implementation to improve
performance and reduce the need to query S3 for recently committed
data.  Utilizes the LRU cache from github.com/karlseguin/ccashe

The logic of the cache is tuned for write-once data:
  * Data is cached with a long (12 hour) expiration.
  * We only cache the presence of data, not it's absence.

We do still manually purge cache keys when data is deleted, so the
only real assumption the cache is making is that our data store won't
be altered by a different actor. This is a fair assumption to make,
even if multiple nodes are sharing a data store, if the data written
is global and static, which should be the case for committed blocks.

It is only recommended to use this version of cache to store records
that are write-once with deterministic keys (for all clients sharing
a single data store). As this matches the intended use-case of S3DB,
we simply replace the registration of S3DBBacked with this
implementation rather than adding a second definition.