Add SelectiveS3DB implementation
SelectiveS3DB combines two databases, matching a regex against the key
to determine which database to use for a given item. This is intended
to allow for certain high-volatility records to be stored in a local
database while storing the larger, read-once records off-box in an S3
bucket. Make note of the `useSlow` function implementation, which
splits data based on regex matching keys.

The current implementation is not substantially faster than CachedS3DB
under real-world conditions and is not recommended unless further
performance testing and/or tuning proves it to be worth the trade-off
of separating the data and increased complexity.