Enhance SocketVal to pre-sign with FilePV
If FilePV is initialized, pass this into the SocketVal initializer.
SocketVal will then attempt to sign w/ the FilePV _before_ sending
the remote sign request, taking advantage of any state management and
double-sign protection that FilePV provides. If the FilePV fails to
sign, that error is returned. Otherwise, the SocketVal leverages the
remote signer as before.

Defense-in-depth solution intended to address
https://github.com/tendermint/tendermint/issues/3328